split dict decode from segment iterator
